feat(#838): migrate 5 task-queue @rpc_expose methods off NexusFS kernel to TaskQueueService
Add @rpc_expose to all 5 TaskQueueService methods (submit_task, get_task,
cancel_task, list_queue_tasks, get_task_stats). Rename list_tasks →
list_queue_tasks for RPC name consistency. Add task_queue_service to
_brick_sources in fastapi_server.py. Delete 5 pure-delegation stubs from
nexus_fs.py.
