+++ src/nexus/core/nexus_fs.py
@@ -10223,71 +10223,6 @@ class NexusFS(  # type: ignore[misc]
             context=context,
         )
 
-    # =========================================================================
-    # TaskQueueService Delegation Methods (5 methods)
-    # Replaces NexusFSTasksMixin (Issue #1387)
-    # =========================================================================
-
-    @rpc_expose(description="Submit a task to the durable task queue")
-    def submit_task(
-        self,
-        task_type: str,
-        params_json: str = "{}",
-        priority: int = 2,
-        max_retries: int = 3,
-        context: OperationContext | None = None,  # noqa: ARG002
-    ) -> dict[str, Any]:
-        """Submit a task to the durable task queue."""
-        return self.task_queue_service.submit_task(
-            task_type=task_type,
-            params_json=params_json,
-            priority=priority,
-            max_retries=max_retries,
-        )
-
-    @rpc_expose(description="Get task status and result")
-    def get_task(
-        self,
-        task_id: int,
-        context: OperationContext | None = None,  # noqa: ARG002
-    ) -> dict[str, Any] | None:
-        """Get task status, progress, and result."""
-        return self.task_queue_service.get_task(task_id)
-
-    @rpc_expose(description="Cancel a pending or running task")
-    def cancel_task(
-        self,
-        task_id: int,
-        context: OperationContext | None = None,  # noqa: ARG002
-    ) -> dict[str, Any]:
-        """Cancel a pending or running task."""
-        return self.task_queue_service.cancel_task(task_id)
-
-    @rpc_expose(description="List tasks with optional filters")
-    def list_queue_tasks(
-        self,
-        task_type: str | None = None,
-        status: int | None = None,
-        limit: int = 50,
-        offset: int = 0,
-        context: OperationContext | None = None,  # noqa: ARG002
-    ) -> list[dict[str, Any]]:  # type: ignore[valid-type]
-        """List tasks with optional filters."""
-        return self.task_queue_service.list_tasks(
-            task_type=task_type,
-            status=status,
-            limit=limit,
-            offset=offset,
-        )
-
-    @rpc_expose(description="Get task queue statistics")
-    def get_task_stats(
-        self,
-        context: OperationContext | None = None,  # noqa: ARG002
-    ) -> dict[str, Any]:
-        """Get task queue statistics."""
-        return self.task_queue_service.get_task_stats()
-
     def close(self) -> None:
         """Close the filesystem and release resources."""
         # Stop DeferredPermissionBuffer first to flush pending permissions
